fix: add int conversion for Windows compatibility in readPassword
On Windows, syscall.Stdin is of type syscall.Handle (uintptr), while on
Unix systems it's an int. The term.ReadPassword function expects an int,
so we need to cast it. Add nolint directive to suppress unconvert warning
on Unix platforms where the cast is unnecessary but harmless.

Fixes Windows CI build error:
  cannot use syscall.Stdin (variable of uintptr type syscall.Handle)
  as int value in argument to term.ReadPassword